Settings UI: move Connections to dashboard. Updated NavigationSettings to pass.
