+++ _inc/client/at-a-glance/index.jsx
@@ -18,6 +18,7 @@ import DashAkismet from './akismet';
 import DashBackups from './backups';
 import DashPluginUpdates from './plugins';
 import DashPhoton from './photon';
+import DashConnections from './connections';
 import { isModuleActivated as _isModuleActivated } from 'state/modules';
 import QuerySitePlugins from 'components/data/query-site-plugins';
 import QuerySite from 'components/data/query-site';
@@ -48,10 +49,6 @@ const AtAGlance = React.createClass( {
 						: 'https://wordpress.com/settings/security/' + this.props.siteRawUrl
 					}
 					externalLinkClick={ () => analytics.tracks.recordEvent( 'jetpack_wpa_aag_security_wpcom_click', {} ) }
-				/>,
-			performanceHeader =
-				<DashSectionHeader
-					label={ __( 'Performance' ) }
 				/>;
 
 		// If user can manage modules, we're in an admin view, otherwise it's a non-admin view.
@@ -92,13 +89,23 @@ const AtAGlance = React.createClass( {
 					</div>
 
 					{
-						// Performance
-						performanceHeader
+						<DashSectionHeader
+							label={ __( 'Performance' ) }
+						/>
 					}
 					<div className="jp-at-a-glance__item-grid">
 						<div className="jp-at-a-glance__left">
 							<DashPhoton />
 						</div>
+					</div>
+
+					{
+						<DashSectionHeader
+							label={ __( 'Connections' ) }
+						/>
+					}
+					<div className="jp-at-a-glance__item-grid">
+						<DashConnections />
 					</div>
 				</div>
 			);
