+++ _inc/client/at-a-glance/style.scss
@@ -283,3 +283,99 @@
 .dops-chart__tooltip .tip-arrow {
 	display: none;
 }
+
+// Connection Settings
+.jp-connections, .jp-connection-settings__info {
+	display: flex;
+	width: 100%;
+}
+.jp-connection-type {
+	width: 50%;
+	flex-grow: 1;
+	display: flex;
+	flex-direction: column;
+
+	&:first-of-type {
+		margin-right: rem( 16px );
+	}
+
+	& > .dops-section-header {
+		width: 100%;
+	}
+
+	& > .dops-card:last-child {
+		flex-grow: 1;
+		display: flex;
+		width: 100%;
+	}
+
+	.jp-connection-settings__actions {
+		margin: 1em 0 0;
+
+		a {
+			cursor: pointer;
+		}
+	}
+}
+@include breakpoint( "<660px" ) {
+	.jp-connections {
+		display: block;
+	}
+	.jp-connection-type {
+		width: 100%;
+	}
+}
+.jp-connection-settings__text {
+	width: 72%;
+	margin-left: rem( 16px );
+}
+.jp-connection-settings__info {
+	.gridicon {
+		opacity: .6;
+	}
+	.gridicon,
+	.jp-connection-settings__site-icon {
+		background: #c8d7e1;
+		color: #fff;
+	}
+	.jp-connection-settings__gravatar {
+		margin-bottom: 0;
+	}
+}
+
+.jp-connection-settings__gravatar {
+	display: inline-block;
+	margin-bottom: rem( 16px );
+	width: rem( 64px );
+	height: rem( 64px );
+	background: $gray;
+	border-radius: 50%;
+}
+
+.jp-connection-settings__username {
+	font-weight: 600;
+}
+
+.jp-connection-settings__email {
+	color: $gray;
+	font-size: rem( 13px );
+	font-style: italic;
+	font-weight: 400;
+}
+
+.jp-connection-settings__cancel {
+	margin-right: 1em;
+}
+
+.jp-connection-settings__modal-body {
+	margin: 0;
+}
+
+.jp-connection-settings__modal-actions {
+	text-align: right;
+	margin-top: 1em;
+}
+
+.jp-connection-settings__cancel, .jp-connection-settings__disconnect {
+	vertical-align: middle;
+}
